+++ PE1/BinomialBinaryOperatorNode.h
@@ -34,6 +34,6 @@ public:
 		}
 	}
 
-	float Extract(const float& x) override;
+	double Extract(const double& x) override;
 };
