+++ PE1/BinomialZeronaryOperationNode.h
@@ -4,13 +4,13 @@
 class BinomialZeronaryOperationNode : public BinomialNode
 {
 private: 
-	float _numeral;
+	double _numeral;
 
 public:
 	BinomialZeronaryOperationNode(StringParser& str, Subset set) : BinomialNode(str, set) {
 		_numeral = str.ParseNumberAtPosition(set);
 	}
 
-	float Extract(const float& x) override;
+	double Extract(const double& x) override;
 };
